UWP: Got a manual animation working (via Step button)
Discovered that the live visual tree for XAML only includes objects that are defined in .xaml files, not dynamically created ones. Added test code to create a circle and rectangle and animate them with the Step button. This works as expected, there must be something happening with timers. Next thing to try is manual stepping through scenarios with breakpoints.